add:user_playlist_detail_del

+++ app/Models/Playlist.php
@@ -192,20 +192,30 @@ class Playlist extends Model
             if(!$data['detail_id'])    return ['id' => null, 'error_code' => 2];   //データ不足
             
             make_error_log("chgPlaylist_detail.log","delete_pl_id=".$data['id']);
-            switch($data['fnc']){
-                case "add":
-                    $pl_id = DB::table('playlistdetail')->insert(['pl_id' => $data['id'],'mus_id' => $data['detail_id']]);
-                    break;
-                case "del":
-                    $deletedRows = DB::table('playlistdetail')->where(['pl_id'=>$data['id'],'mus_id'=>$data['detail_id']])->delete();
-                    if ($deletedRows > 0) return ['id' => null, 'error_code' => 0];   //成功
-                    else return ['id' => null, 'error_code' => -1];   //失敗
-                    break;
-                default:
+
+            //管理者以外は自身のﾌﾟﾚｲﾘｽﾄのみ更新可能
+            $exists = null;
+            $user = Auth::user();
+            if(!$user->admin_flag){
+                $exists = Playlist::where('id', $data['id'])->where('user_id', $user->id)->exists();
+                if(!$exists) 
                     return ['id' => null, 'error_code' => -1];   //失敗
-                    break;
-                    
             }
+            if($data['fnc'] == "add"){
+                $pl_id = DB::table('playlistdetail')->insert(['pl_id' => $data['id'],'mus_id' => $data['detail_id']]);
+
+            }elseif($data['fnc'] == "del"){
+                $deletedRows = DB::table('playlistdetail')->where(['pl_id'=>$data['id'],'mus_id'=>$data['detail_id']])->delete();
+                if ($deletedRows > 0) 
+                    return ['id' => null, 'error_code' => 0];   //成功
+                else 
+                    return ['id' => null, 'error_code' => -1];   //失敗
+
+            }else{
+                return ['id' => null, 'error_code' => -1];   //失敗
+            }
+
+
             make_error_log("chgPlaylist_detail.log","success");
             return ['id' => null, 'error_code' => 0];   //成功
         } catch (\Exception $e) {
